Changes needed for pedigree submission
There will be corresponding changes in cgap-portal to enable
submission of pedigree metadata as well as accessioning metadata.
These are the changes needed on the SubmitCGAP side to work with
those cgap-portal changes. Not specifying a submission_type in the
command should default to 'accessioning', so it should still work with
older command structure.